chore(checker): keep TS2344 constraint helper under arch guard
